Rename noprovider pkg to genericprovider
