+++ internal/providers/default.go
@@ -3,10 +3,10 @@
 package providers
 
 import (
-	"github.com/ubuntu/authd-oidc-brokers/internal/providers/noprovider"
+	"github.com/ubuntu/authd-oidc-brokers/internal/providers/genericprovider"
 )
 
 // CurrentProvider returns a generic oidc provider implementation.
 func CurrentProvider() Provider {
-	return noprovider.New()
+	return genericprovider.New()
 }
